Reworked logging screen:   - Autocomplete for categories   - Autofill for ABV when choosing a pre made drink   - Cleaner UI design
TODO:
  - Add recipient field
  - Add time and date field

+++ app/src/main/java/com/example/alcoholtracker/data/model/UserDrinkLogSummary.kt
@@ -3,5 +3,5 @@ package com.example.alcoholtracker.data.model
 data class UserDrinkLogSummary(
     var totalCost: Double,
     var drinkCount: Int,
-    var totalAmount: Double
+    var totalAmount: Int
 )
